[i386] Change [u]int32_t to [unsigned] int, rather than [unsigned] long
This brings us in to line with Linux definitions, and also simplifies
adding x86_64 support since both platforms have 2-byte shorts, 4-byte
ints and 8-byte long longs.

+++ src/arch/i386/include/gdbmach.h
@@ -12,7 +12,7 @@
 
 #include <stdint.h>
 
-typedef uint32_t gdbreg_t;
+typedef unsigned long gdbreg_t;
 
 /* The register snapshot, this must be in sync with interrupt handler and the
  * GDB protocol. */
